// frontend/src/generated/types/consumerOrderCreateRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { ConsumerOrderCreateItemRequest } from './consumerOrderCreateItemRequest';

export interface ConsumerOrderCreateRequest {
  /** @maxLength 36 */
  clientRequestId: string;
  /** @minLength 1 */
  items: ConsumerOrderCreateItemRequest[];
  /** MVP에서는 null 또는 공백만 허용 */
  requestNote?: string;
}